requirements.txt
annotated-types==0.6.0
anyio==3.7.1
black==23.11.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-speech==2.21.0
google-cloud-storage==2.9.0
google-cloud-translate==3.12.1
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
idna==3.4
isort==5.12.0
Levenshtein==0.23.0
mypy-extensions==1.0.0
nodeenv==1.8.0
openai==1.3.3
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pydantic==2.5.1
pydantic_core==2.14.3
pyparsing==3.1.1
pyright==1.1.336
rapidfuzz==3.5.2
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
